+++ CHANGELOG.md
@@ -62,6 +62,8 @@
   [#4103](https://github.com/Kong/kong-operator/pull/4103)
 - Hybridgateway: add support for `konghq.com/retries` service annotation
   [#4104](https://github.com/Kong/kong-operator/pull/4104)
+- Hybridgateway: add support for `konghq.com/host-header` service annotation
+  [#4108](https://github.com/Kong/kong-operator/pull/4108)
 - Add the following headers in requests of `ingress-controller` sent to Konnect
   for uploading configuration for tracing:
   - `X-Kic-Konnect-Sync-Instance-Id` for instance ID of Konnect config synchronizer.